pyproject.toml
[project]
name = "projetoIvo"
version = "0.1.0"
description = "projeto do ivo e fabio"
authors = [
    {name = "wtfemanu", email = "os9777190@gmail.com"},
]
dependencies = ["django>=5.2.3", "djangorestframework>=3.16.0"]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
